test/mochitest/test_check_adbhelper.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
  <div class="dialog error" id="errorDisplay">
    <section class="unknown">
      An unknown error has been reported. Please check WebConsole and JSConsole.
    </section>
    <section class="adbhelper">
      Please verify ADB Helper v0.8.6 is installed and enabled.
    </section>
    <section class="gecko-tooold">
      Your Firefox Nightly is too old, please update.
    </section>
    <section class="gecko-unsupported-os-arch">
      This addon is only targetting Firefix Nightly running on Linux 32/64 bits or OSX 64 bits.
    </section>
    <section class="gecko-unsupported-build">
      This addon is only targetting Firefox Nightly builds.
    </section>
  </div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  function test_verify_adbhelper_good() {
    AddonManager = {
      getAddonByID: function(aID, aCallback) {
        aCallback({ userDisabled: false, version: "0.8.6" });
        return;
      }
    };

    return new Promise((resolve, reject) => {
      verifyADBHelper().then(() => {
        ok(true, "ADBHelper verification success");
        resolve();
      });
    });
  }

  function test_verify_adbhelper_missing() {
    AddonManager = {
      getAddonByID: function(aID, aCallback) {
        aCallback(null);
        return;
      }
    };

    return new Promise((resolve, reject) => {
      verifyADBHelper().catch((err) => {
        is(err.message, "ADBHELPER_MISSING", "return missing error");
        resolve();
      });
    });
  }


  function test_verify_adbhelper_disabled() {
    AddonManager = {
      getAddonByID: function(aID, aCallback) {
        aCallback({ userDisabled: true });
        return;
      }
    };

    return new Promise((resolve, reject) => {
      verifyADBHelper().catch((err) => {
        is(err.message, "ADBHELPER_DISABLED", "return disabled error");
        resolve();
      });
    });
  }

  function test_verify_adbhelper_version() {
    AddonManager = {
      getAddonByID: function(aID, aCallback) {
        aCallback({ userDisabled: false, version: "0.8.5" });
        return;
      }
    };

    return new Promise((resolve, reject) => {
      verifyADBHelper().catch((err) => {
        is(err.message, "ADBHELPER_VERSION", "return version error");
        resolve();
      });
    });
  }

  function runTest() {
    loadAboutJs()
      .then(() => test_verify_adbhelper_good())
      .then(() => test_verify_adbhelper_missing())
      .then(() => test_verify_adbhelper_disabled())
      .then(() => test_verify_adbhelper_version())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
